% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab5.R
\docType{class}
\name{kolada-class}
\alias{kolada-class}
\alias{kolada}
\title{Access and analyse the data in the Kolada dataset}
\description{
This RC class function can access to the Kolada dataset which kpi's id is N00945 via Kolada API. Check here: http://api.kolada.se/v2/kpi/N00945.
Ihe data is percentage of temporary parental benefit (in terms of net days) taken out by men from 1996 to 2022.
This function just gets these five cities' data: Stockholm, Gothenburg, Lund, Uppsala, and Linkoping.
}
\section{Fields}{

\describe{
\item{\code{city}}{The city you want to query. You could only choose from Stockholm, Gothenburg, Lund, Uppsala, and Linkoping.}

\item{\code{city_and_id}}{Cities and their ids in Kolada dataset. Check here: https://api.kolada.se/v2/municipality.}

\item{\code{one_city_data}}{Data frame of the city you queried.}

\item{\code{all_city_data}}{Data frame of all the five cities.}
}}


